+++ src/axom/core/utilities/StringUtilities.cpp
@@ -39,6 +39,27 @@ void toUpper(std::string& str)
   });
 }
 
+std::vector<std::string> splitLastNTokens(const std::string& input,
+                                          const std::size_t n,
+                                          const char delim)
+{
+  std::vector<std::string> result;
+
+  auto last_pos = std::string::npos;
+  auto pos = input.find_last_of(delim, last_pos - 1);
+
+  while((pos != std::string::npos) && (result.size() < n - 1))
+  {
+    result.push_back(input.substr(pos + 1, last_pos - pos - 1));
+    last_pos = pos;
+    pos = input.find_last_of(delim, last_pos - 1);
+  }
+  // Add the rest of the string (first token)
+  result.push_back(input.substr(0, last_pos));
+  std::reverse(result.begin(), result.end());
+  return result;
+}
+
 }  // end namespace string
 }  // end namespace utilities
 }  // end namespace axom