+++ src/axom/inlet/Inlet.hpp
@@ -58,14 +58,23 @@ public:
    * \param [in] sidreRootGroup Pointer to the already created Sidre Group.
    * \param [in] docEnabled Boolean indicating whether documentation generation
    * is enabled. This also toggles the storing of documentation-specific information.
+   * \param [in] reconstruct Whether or not to attempt to reconstruct child Containers
+   * and Fields from the data in the sidre Group
    *****************************************************************************
    */
   Inlet(std::unique_ptr<Reader> reader,
         axom::sidre::Group* sidreRootGroup,
-        bool docEnabled = true)
+        bool docEnabled = true,
+        bool reconstruct = false)
     : m_reader(std::move(reader))
     , m_sidreRootGroup(sidreRootGroup)
-    , m_globalContainer("", "", *m_reader, m_sidreRootGroup, m_unexpectedNames, docEnabled)
+    , m_globalContainer("",
+                        "",
+                        *m_reader,
+                        m_sidreRootGroup,
+                        m_unexpectedNames,
+                        docEnabled,
+                        reconstruct)
     , m_docEnabled(docEnabled)
   {
     m_unexpectedNames = m_reader->getAllNames();
@@ -476,7 +485,7 @@ public:
    * in the input file that were not added via an add* call
    *****************************************************************************
    */
-  const std::unordered_set<std::string>& unexpectedNames() const
+  const std::vector<std::string>& unexpectedNames() const
   {
     return m_unexpectedNames;
   }
@@ -488,7 +497,7 @@ private:
   Container m_globalContainer;
   std::unique_ptr<Writer> m_writer;
   bool m_docEnabled;
-  std::unordered_set<std::string> m_unexpectedNames;
+  std::vector<std::string> m_unexpectedNames;
 };
 
 }  // end namespace inlet